@@ -15,10 +15,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 0.7.5+20
+version: 0.7.5+22
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.11.0 <3.0.0'
 
 dependencies:
   auto_route: ^0.6.9
@@ -34,8 +34,10 @@ dependencies:
   provider: ^4.3.2+2
   shared_preferences: '>=0.5.12+4 <2.0.0'
   stacked: ^1.7.6
-  stacked_services: ^0.5.4+5
+  stacked_services: ^0.6.0
   url_launcher: ^5.7.10
+  
+  shake: ^0.1.0
   
   spotify: ^0.4.0
   spotify_sdk: ^1.0.0
@@ -56,9 +58,10 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.1+1
 
 dev_dependencies:
+  pedantic: ^1.9.2
   auto_route_generator:
   injectable_generator:
   build_runner:
