@@ -1,4 +1,4 @@
 class SpotifyManager {
-  static const clientId = String.fromEnvironment("SPOTIFY_CLIENT_ID");
-  static const clientSecret = String.fromEnvironment("SPOTIFY_CLIENT_SECRET");
+  static const clientId = String.fromEnvironment('SPOTIFY_CLIENT_ID');
+  static const clientSecret = String.fromEnvironment('SPOTIFY_CLIENT_SECRET');
 }
