@@ -11,14 +11,13 @@ class TimerServiceManager {
     return _instance;
   }
 
-  Map<String, TimerService> _timerServices = {};
+  final Map<String, TimerService> _timerServices = {};
 
   TimerService getTimerService(final String id) {
     return _timerServices[id];
   }
 
-  void setTimerService(
-      final TimerService timerService) {
+  void setTimerService(final TimerService timerService) {
     _timerServices[timerService.timerModel.id] = timerService;
   }
 }
