@@ -4,18 +4,18 @@ class App {
   String packageName;
 
   Map<dynamic, dynamic> toMap() {
-    final Map<dynamic, dynamic> pigeonMap = <dynamic, dynamic>{};
-    pigeonMap['title'] = title;
-    pigeonMap['icon'] = icon;
-    pigeonMap['packageName'] = packageName;
+    final pigeonMap = <dynamic, dynamic>{
+      'title': title,
+      'icon': icon,
+      'packageName': packageName,
+    };
     return pigeonMap;
   }
 
   static App fromMap(Map<dynamic, dynamic> pigeonMap) {
-    final App result = App();
-    result.title = pigeonMap['title'];
-    result.icon = pigeonMap['icon'];
-    result.packageName = pigeonMap['packageName'];
+    final result = App().title = pigeonMap['title'].icon =
+        pigeonMap['icon'].packageName = pigeonMap['packageName'];
+
     return result;
   }
 }
